requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.5.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cuda-python==11.8.0
cycler==0.12.1
Cython==3.0.11
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.44.1
fsspec==2023.10.0
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
kiwisolver==1.4.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.3
mdurl==0.1.2
mkl-service==2.4.0
mpmath==1.3.0
onnx==1.16.2
onnxruntime-gpu==1.18.0
onnxsim==0.4.36
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
platformdirs==4.0.0
pluggy==1.5.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pycuda==2024.1.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.1
pytest==8.3.3
python-dateutil==2.8.2
python-multipart==0.0.10
pytools==2024.1.14
pytz==2024.2
PyYAML==6.0.1
regex==2023.10.3
rich==13.8.1
safetensors==0.4.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
supervision==0.6.0
sympy==1.13.1
timm==0.9.10
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.102
ultralytics-thop==2.0.8
uvicorn==0.30.6
yapf==0.40.2
zipp==3.17.0
